+++ cordapp-cpk2/src/main/kotlin/net/corda/plugins/cpk2/CPKDependenciesTask.kt
@@ -28,7 +28,7 @@ import java.util.Base64
 import java.util.jar.JarFile
 import javax.inject.Inject
 
-@Suppress("UnstableApiUsage", "MemberVisibilityCanBePrivate")
+@Suppress("MemberVisibilityCanBePrivate")
 @DisableCachingByDefault
 open class CPKDependenciesTask @Inject constructor(objects: ObjectFactory) : DefaultTask() {
     private companion object {