+++ cordapp-cpk2/src/test/resources/cordapp-with-guava/src/main/java/com/example/contract/GuavaContract.java
@@ -3,12 +3,13 @@ package com.example.contract;
 import com.google.common.collect.ImmutableList;
 import net.corda.v5.ledger.contracts.Contract;
 import net.corda.v5.ledger.transactions.LedgerTransaction;
+import org.jetbrains.annotations.NotNull;
 
 import java.util.List;
 
 public class GuavaContract implements Contract {
     @Override
-    public void verify(LedgerTransaction ltx) {
+    public void verify(@NotNull LedgerTransaction ltx) {
         List<String> data = ImmutableList.of("one", "two", "three");
         for (String item : data) {
             System.out.println(item);